requirements.txt
pybit>=2.6.0
python-dotenv>=1.0.1
tenacity>=8.2.2